requirements.txt
asn1crypto==1.5.1
attrs==22.1.0
build==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
exceptiongroup==1.0.4
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
more-itertools==9.0.0
packaging==21.3
pep517==0.13.0
pg8000==1.29.3
pluggy==0.13.1
psycopg2==2.9.5
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scramp==1.4.4
six==1.16.0
SQLAlchemy==1.4.44
swagger-ui-bundle==0.0.9
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
